refactor: renaming date-picker component to date-picker-calendar
