+++ .codeclimate.yml
@@ -1,18 +1,18 @@
 exclude_patterns:
-  - "../src/test.ts"
-  - "../src/**/*.spec.ts"
-  - "../projects/**/*.spec.ts"
-  - "config/"
-  - "db/"
-  - "dist/"
-  - "features/"
-  - "**/node_modules/"
-  - "script/"
-  - "**/spec/"
-  - "**/test/"
-  - "**/tests/"
-  - "Tests/"
-  - "**/vendor/"
-  - "**/*_test.go"
-  - "**/*.d.ts"
-  - "**/*.spec.ts"
+  - '../src/test.ts'
+  - '../src/**/*.spec.ts'
+  - '../projects/**/*.spec.ts'
+  - 'config/'
+  - 'db/'
+  - 'dist/'
+  - 'features/'
+  - '**/node_modules/'
+  - 'script/'
+  - '**/spec/'
+  - '**/test/'
+  - '**/tests/'
+  - 'Tests/'
+  - '**/vendor/'
+  - '**/*_test.go'
+  - '**/*.d.ts'
+  - '**/*.spec.ts'
