+++ projects/ion/src/lib/picker/date-picker/date-picker-calendar/date-picker-calendar.component.scss
